style(lint): validate project with swiftlint
