+++ .swiftlint.yml
@@ -1,7 +1,9 @@
 disabled_rules: # rule identifiers to exclude from running
   - function_body_length
   - line_length
+  - file_length
   - identifier_name
   - empty_enum_arguments
+  - force_try
 excluded: # paths to ignore during linting. Takes precedence over `included`.
   - Pods
