+++ flyve-mdm-ios/AppDelegate.swift
@@ -30,10 +30,10 @@ import CoreLocation
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
-    
+
     var window: UIWindow?
     let manager = CLLocationManager()
-    
+
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
 
         if CLLocationManager.locationServicesEnabled() {
@@ -42,94 +42,85 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                 manager.requestWhenInUseAuthorization()
             }
         }
-        
+
         if let mdmAgentData = getStorage(key: "mdmAgent") as? [String: AnyObject] {
-            
+
             loadMainView(userToken: "", invitationToken: "", mdmAgent: mdmAgentData)
-            
+
         } else {
             loadMainView(userToken: "", invitationToken: "")
         }
-    
+
         return true
     }
-    
+
     func loadMainView(userToken: String, invitationToken: String, mdmAgent: [String: Any] = [String: Any]()) {
-        
+
         window = UIWindow(frame: UIScreen.main.bounds)
         window?.makeKeyAndVisible()
-        
+
         if mdmAgent.count > 0 {
             window?.rootViewController = UINavigationController(rootViewController: MainController(mdmAgent: mdmAgent))
         } else {
             window?.rootViewController = UINavigationController(rootViewController: ViewController(userToken: userToken, invitationToken: invitationToken))
         }
     }
-    
+
     func application(_ application: UIApplication, open url: URL, sourceApplication: String?, annotation: Any) -> Bool {
-        
+
         if let mdmAgentData = getStorage(key: "mdmAgent") as? [String: AnyObject] {
-            
             loadMainView(userToken: "", invitationToken: "", mdmAgent: mdmAgentData)
-            
         }
-        
+
         var invitation = [String: AnyObject]()
-        
         guard let urlEnroll = URLComponents(string: url.absoluteString) else {
             loadMainView(userToken: "", invitationToken: "")
             return true
         }
-    
+
         guard let query = urlEnroll.queryItems?.first(where: { $0.name == "data" })?.value else {
-            
             loadMainView(userToken: "", invitationToken: "")
             return true
         }
-        
+
         if let data = query.base64Decoded()?.data(using: .utf8) {
             do {
-                invitation =  try JSONSerialization.jsonObject(with: data, options: []) as! [String: AnyObject]
+                invitation =  try JSONSerialization.jsonObject(with: data, options: []) as? [String: AnyObject] ?? [String: AnyObject]()
             } catch {
                 print(error.localizedDescription)
             }
         }
-        
+
         guard let url_invitation: String = invitation["url"] as? String, let user_token: String = invitation["user_token"] as? String, let invitation_token: String = invitation["invitation_token"] as? String else {
-            
             loadMainView(userToken: "", invitationToken: "")
             return true
         }
-        
-        baseURL = url_invitation
 
+        baseURL = url_invitation
         loadMainView(userToken: user_token, invitationToken: invitation_token)
-        
+
         return true
     }
-    
+
     func applicationWillResignActive(_ application: UIApplication) {
         // Sent when the application is about to move from active to inactive state. This can occur for certain types of temporary interruptions (such as an incoming phone call or SMS message) or when the user quits the application and it begins the transition to the background state.
         // Use this method to pause ongoing tasks, disable timers, and invalidate graphics rendering callbacks. Games should use this method to pause the game.
     }
-    
+
     func applicationDidEnterBackground(_ application: UIApplication) {
         // Use this method to release shared resources, save user data, invalidate timers, and store enough application state information to restore your application to its current state in case it is terminated later.
         // If your application supports background execution, this method is called instead of applicationWillTerminate: when the user quits.
     }
-    
+
     func applicationWillEnterForeground(_ application: UIApplication) {
         // Called as part of the transition from the background to the active state; here you can undo many of the changes made on entering the background.
     }
-    
+
     func applicationDidBecomeActive(_ application: UIApplication) {
         // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
     }
-    
+
     func applicationWillTerminate(_ application: UIApplication) {
         // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
     }
-    
-    
 }
-
