+++ flyve-mdm-ios/Delay+Helper.swift
@@ -27,7 +27,6 @@
 
 import Foundation
 
-public func delay(_ delay: Double = 3.0, closure: @escaping ()->()) {
-    DispatchQueue.main.asyncAfter(
-        deadline: DispatchTime.now() + Double(Int64(delay * Double(NSEC_PER_SEC))) / Double(NSEC_PER_SEC), execute: closure)
+public func delay(_ delay: Double = 3.0, closure: @escaping () -> Void) {
+    DispatchQueue.main.asyncAfter(deadline: DispatchTime.now() + Double(Int64(delay * Double(NSEC_PER_SEC))) / Double(NSEC_PER_SEC), execute: closure)
 }
