+++ flyve-mdm-ios/FieldInfoCell.swift
@@ -1,0 +1,76 @@
+/*
+ *   Copyright © 2017 Teclib. All rights reserved.
+ *
+ * FieldInfoCell.swift is part of flyve-mdm-ios
+ *
+ * flyve-mdm-ios is a subproject of Flyve MDM. Flyve MDM is a mobile
+ * device management software.
+ *
+ * flyve-mdm-ios is free software: you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 3
+ * of the License, or (at your option) any later version.
+ *
+ * flyve-mdm-ios is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * ------------------------------------------------------------------------------
+ * @author    Hector Rondon
+ * @date      24/07/17
+ * @copyright Copyright © 2017 Teclib. All rights reserved.
+ * @license   GPLv3 https://www.gnu.org/licenses/gpl-3.0.html
+ * @link      https://github.com/flyve-mdm/flyve-mdm-ios
+ * @link      https://flyve-mdm.com
+ * ------------------------------------------------------------------------------
+ */
+
+import UIKit
+
+class FieldInfoCell: UITableViewCell {
+
+    override init(style: UITableViewCellStyle, reuseIdentifier: String!) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+
+        self.selectionStyle = UITableViewCellSelectionStyle.none
+        self.contentView.backgroundColor = .clear
+        self.setupView()
+        self.addConstraints()
+
+    }
+
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+    }
+
+    func setupView() {
+
+        backgroundColor = .white
+        contentView.addSubview(self.textField)
+    }
+
+    func addConstraints() {
+
+        self.textField.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: 8).isActive = true
+        self.textField.bottomAnchor.constraint(equalTo: self.contentView.bottomAnchor, constant: -8).isActive = true
+        self.textField.leadingAnchor.constraint(equalTo: self.contentView.leadingAnchor, constant: 8).isActive = true
+        self.textField.trailingAnchor.constraint(equalTo: self.contentView.trailingAnchor, constant: -8).isActive = true
+    }
+
+    let textField: UITextField = {
+
+        let text = UITextField()
+        text.translatesAutoresizingMaskIntoConstraints = false
+        text.placeholder = ""
+        text.clearButtonMode = UITextFieldViewMode.whileEditing
+        text.textColor = .gray
+        text.keyboardType = .default
+        text.borderStyle = .none
+
+        return text
+    }()
+}
