+++ flyve-mdm-ios/TitleInfoCell.swift
@@ -1,0 +1,87 @@
+/*
+ *   Copyright © 2017 Teclib. All rights reserved.
+ *
+ * TitleInfoCell.swift is part of flyve-mdm-ios
+ *
+ * flyve-mdm-ios is a subproject of Flyve MDM. Flyve MDM is a mobile
+ * device management software.
+ *
+ * flyve-mdm-ios is free software: you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 3
+ * of the License, or (at your option) any later version.
+ *
+ * flyve-mdm-ios is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * ------------------------------------------------------------------------------
+ * @author    Hector Rondon
+ * @date      24/07/17
+ * @copyright Copyright © 2017 Teclib. All rights reserved.
+ * @license   GPLv3 https://www.gnu.org/licenses/gpl-3.0.html
+ * @link      https://github.com/flyve-mdm/flyve-mdm-ios
+ * @link      https://flyve-mdm.com
+ * ------------------------------------------------------------------------------
+ */
+
+import UIKit
+
+class TitleInfoCell: UITableViewCell {
+
+    override init(style: UITableViewCellStyle, reuseIdentifier: String!) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+
+        self.selectionStyle = UITableViewCellSelectionStyle.none
+        self.contentView.backgroundColor = .clear
+        self.setupView()
+    }
+
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+    }
+
+    func setupView() {
+
+        backgroundColor = .white
+        contentView.addSubview(self.titleLabel)
+        contentView.addSubview(self.lineView)
+
+        addConstraints()
+    }
+
+    func addConstraints() {
+
+        self.lineView.heightAnchor.constraint(equalToConstant: 0.5).isActive = true
+        self.lineView.leftAnchor.constraint(equalTo: self.contentView.leftAnchor).isActive = true
+        self.lineView.rightAnchor.constraint(equalTo: self.contentView.rightAnchor).isActive = true
+
+        self.titleLabel.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: 8).isActive = true
+        self.titleLabel.bottomAnchor.constraint(equalTo: self.lineView.topAnchor, constant: -8).isActive = true
+        self.titleLabel.leftAnchor.constraint(equalTo: self.contentView.leftAnchor, constant: 8).isActive = true
+        self.titleLabel.rightAnchor.constraint(equalTo: self.contentView.rightAnchor, constant: -8).isActive = true
+    }
+
+    let lineView: UIView = {
+
+        let line = UIView()
+
+        line.translatesAutoresizingMaskIntoConstraints = false
+        line.backgroundColor = .gray
+
+        return line
+    }()
+
+    let titleLabel: UILabel = {
+        let label = UILabel()
+
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.textColor = .gray
+
+        return label
+    }()
+}
